+++ sydtest-sqitch-postgres/test_resources/toy-sqitch-hardcoded-public/deploy/init.sql
@@ -1,0 +1,9 @@
+-- Deploy toy-hardcoded-public:init to pg
+BEGIN;
+
+CREATE TABLE IF NOT EXISTS widget (
+    id   SERIAL PRIMARY KEY,
+    name TEXT NOT NULL
+);
+
+COMMIT;