+++ sydtest-sqitch-postgres/test_resources/toy-sqitch-hardcoded-public/revert/init.sql
@@ -1,0 +1,6 @@
+-- Revert toy-hardcoded-public:init from pg
+BEGIN;
+
+DROP TABLE IF EXISTS widget;
+
+COMMIT;